pyproject.toml
[project]
name = "az-economic-divergence"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dash>=4.1.0",
    "duckdb>=1.5.1",
    "fredapi>=0.5.2",
    "great-expectations>=1.15.2",
    "ipykernel>=7.2.0",
    "pandas>=3.0.2",
    "plotly>=6.6.0",
    "poethepoet>=0.44.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "ruff>=0.15.10",
    "wbdata>=1.1.0",
    "weo>=0.7.4",
]

[tool.poe.tasks]
format = "uv run ruff format"
fix    = "uv run ruff check --fix"

[tool.ruff]
select = ["E", "F", "I"]
